src/api/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pymongo==4.6.2
pytz==2024.1
sqlparse==0.4.4
django-cors-headers==4.3.0
requests==2.31.0
